feat: move menu types to separate file
