+++ packages/app/src/systems/Sidebar/components/Menu/Menu.stories.tsx
@@ -1,6 +1,6 @@
 import { Box } from '@fuel-ui/react';
 
-import type { MenuItemObj, MenuProps } from './Menu';
+import type { MenuItemObj, MenuProps } from './types';
 import { Menu } from './Menu';
 
 export default {
